add refresh glowIndicator behaviour
